Extract `GuestRendererPreload` out of `cuttlefish_config.h` into a new file
Bug: b/436338992
