+++ base/cvd/cuttlefish/host/commands/assemble_cvd/BUILD.bazel
@@ -333,6 +333,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:config_constants",
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:guest_hwui_renderer",
+        "//cuttlefish/host/libs/config:guest_renderer_preload",
         "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
         "@fmt",
