+++ base/cvd/cuttlefish/host/libs/config/BUILD.bazel
@@ -107,6 +107,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:config_utils",
         "//cuttlefish/host/libs/config:external_network_mode",
         "//cuttlefish/host/libs/config:guest_hwui_renderer",
+        "//cuttlefish/host/libs/config:guest_renderer_preload",
         "//cuttlefish/host/libs/config:secure_hals",
         "//cuttlefish/host/libs/config:vmm_mode",
         "//libbase",
@@ -207,6 +208,16 @@ cf_cc_library(
     hdrs = ["guest_os.h"],
     deps = [
         "//cuttlefish/host/libs/config:boot_flow",
+    ],
+)
+
+cf_cc_library(
+    name = "guest_renderer_preload",
+    srcs = ["guest_renderer_preload.cc"],
+    hdrs = ["guest_renderer_preload.h"],
+    deps = [
+        "//cuttlefish/common/libs/utils:result",
+        "//libbase",
     ],
 )
