+++ base/cvd/cuttlefish/host/libs/vm_manager/BUILD.bazel
@@ -47,6 +47,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/host/libs/config:external_network_mode",
         "//cuttlefish/host/libs/config:guest_hwui_renderer",
+        "//cuttlefish/host/libs/config:guest_renderer_preload",
         "//cuttlefish/host/libs/config:known_paths",
         "//cuttlefish/host/libs/config:vmm_mode",
         "//cuttlefish/host/libs/feature",
